Bump Go and rancher/shepherd version and update code (#111)
* Version bump Go to 1.21 and rancher/shepherd


Update other dependencies


Fix failing bump


* Modify GKE provisioning tests


* Modify AKS provisioning tests


* Modify EKS provisioning tests


* Modify GKE import tests


* Attempt at parallelizing support matrix and gke P1 tests


Attempt at fixing P1 tests


* Fix AKS provisioning tests, reorder GKE functions and log clusterName


Attempt at fixing import test cases


Delete clusters only if cluster!=nil


Remove extra spaces


* Add --wait when deleting EKS resources and remove parallelization from support matrix and P1 tests


Remove --disable-nodegroup-eviction from eks delete


Remove programmatic focus and add --force to eks delete


Parallelize support matrix tests for GKE and AKS


---------

+++ hosted/aks/k8s_chart_support/k8s_chart_support_provisioning_test.go
@@ -17,14 +17,14 @@ var _ = Describe("K8sChartSupportProvisioning", func() {
 	)
 	BeforeEach(func() {
 		var err error
-		cluster, err = helper.CreateAKSHostedCluster(ctx.RancherAdminClient, ctx.CloudCred.ID, clusterName, k8sVersion, location, helpers.GetCommonMetadataLabels())
+		cluster, err = helper.CreateAKSHostedCluster(ctx.RancherAdminClient, clusterName, ctx.CloudCred.ID, k8sVersion, location)
 		Expect(err).To(BeNil())
 		cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherAdminClient)
 		Expect(err).To(BeNil())
 	})
 
 	AfterEach(func() {
-		if ctx.ClusterCleanup {
+		if ctx.ClusterCleanup && cluster != nil {
 			err := helper.DeleteAKSHostCluster(cluster, ctx.RancherAdminClient)
 			Expect(err).To(BeNil())
 		} else {
