+++ hosted/aks/k8s_chart_support/upgrade/k8s_chart_support_import_upgrade_test.go
@@ -25,7 +25,7 @@ var _ = Describe("K8sChartSupportUpgradeImport", func() {
 		cluster.AKSConfig = cluster.AKSStatus.UpstreamSpec
 	})
 	AfterEach(func() {
-		if ctx.ClusterCleanup {
+		if ctx.ClusterCleanup && cluster != nil {
 			err := helper.DeleteAKSHostCluster(cluster, ctx.RancherAdminClient)
 			Expect(err).To(BeNil())
 			err = helper.DeleteAKSClusteronAzure(clusterName)
