+++ hosted/aks/p0/p0_importing_test.go
@@ -53,7 +53,7 @@ var _ = Describe("P0Importing", func() {
 			BeforeEach(func() {
 				k8sVersion, err := helper.GetK8sVersion(ctx.RancherAdminClient, ctx.CloudCred.ID, location, testData.isUpgrade)
 				Expect(err).NotTo(HaveOccurred())
-				GinkgoLogr.Info("Using K8s version: " + k8sVersion)
+				GinkgoLogr.Info(fmt.Sprintf("Using K8s version %s for cluster %s", k8sVersion, clusterName))
 
 				err = helper.CreateAKSClusterOnAzure(location, clusterName, k8sVersion, "1", helpers.GetCommonMetadataLabels())
 				Expect(err).To(BeNil())
@@ -67,7 +67,7 @@ var _ = Describe("P0Importing", func() {
 			})
 
 			AfterEach(func() {
-				if ctx.ClusterCleanup {
+				if ctx.ClusterCleanup && cluster != nil {
 					err := helper.DeleteAKSHostCluster(cluster, ctx.RancherAdminClient)
 					Expect(err).To(BeNil())
 					err = helper.DeleteAKSClusteronAzure(clusterName)
