+++ hosted/eks/support_matrix/support_matrix_provisioning_test.go
@@ -40,14 +40,14 @@ var _ = Describe("SupportMatrixProvisioning", func() {
 			BeforeEach(func() {
 				clusterName = namegen.AppendRandomString(helpers.ClusterNamePrefix)
 				var err error
-				cluster, err = helper.CreateEKSHostedCluster(ctx.StdUserClient, clusterName, ctx.CloudCred.ID, version, region, helpers.GetCommonMetadataLabels())
+				cluster, err = helper.CreateEKSHostedCluster(ctx.StdUserClient, clusterName, ctx.CloudCred.ID, version, region)
 				Expect(err).To(BeNil())
 				// Requires RancherAdminClient
 				cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherAdminClient)
 				Expect(err).To(BeNil())
 			})
 			AfterEach(func() {
-				if ctx.ClusterCleanup {
+				if ctx.ClusterCleanup && cluster != nil {
 					err := helper.DeleteEKSHostCluster(cluster, ctx.StdUserClient)
 					Expect(err).To(BeNil())
 				} else {
